hotfix

+++ core/controllers/league.php
@@ -33,7 +33,8 @@
 						$tempname = $_FILES['imageleague']['tmp_name'];    
 						$folder   = "assets/img/leagues/{$leagueId}";
 
-						unlink("../../{$folder}/{$filename}");
+						if(file_exists("../../{$folder}/{$filename}"))
+							unlink("../../{$folder}/{$filename}");
 	          
 	          			mkdir(dirname(__FILE__, 3) . "/{$folder}", 0777, true);          
 						if (move_uploaded_file($tempname, "../../{$folder}/{$filename}"))
