+++ core/controllers/team.php
@@ -32,7 +32,8 @@
 						$tempname = $_FILES['imageteam']['tmp_name'];    
 						$folder   = "assets/img/teams/{$teamId}";
 
-						unlink("../../{$folder}/{$filename}"); 
+						if(file_exists("../../{$folder}/{$filename}"))
+							unlink("../../{$folder}/{$filename}");
 	          
 	          			mkdir(dirname(__FILE__, 3) . "/{$folder}", 0777, true);          
 						if (move_uploaded_file($tempname, "../../{$folder}/{$filename}"))
